Add files via upload
Alternate to the helper scripts - edit_genome.sh and scoring.R - it uses the same infrastructure but it operates on a whole-genome level and prunes the chromosomes that hit the desired similarity score. It was an attempt to decrease runtime, but unfortunately it still requires a lot of trials. An option I have yet to try out is to increase the window within which the score should fall.